+++ packages/config/.eslintrc.cjs
@@ -1,0 +1,8 @@
+module.exports = {
+  root: false,
+  extends: ['../../.eslintrc.cjs'],
+  parserOptions: {
+    project: ['./tsconfig.json'],
+    tsconfigRootDir: __dirname,
+  },
+};